Chapter 07/requirements.txt
streamlit
qdrant-client
llama-index
langchain
langchain-community
langchain-nomic
sentence-transformers
transformers
pillow
pypdf
scikit-learn
numpy
